+++ core/src/main/scala/org/locationtech/rasterframes/expressions/transformers/URIToRasterSource.scala
@@ -27,12 +27,10 @@ import org.locationtech.rasterframes.RasterSourceType
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
 import org.apache.spark.sql.catalyst.expressions.{ExpectsInputTypes, Expression, UnaryExpression}
-import org.apache.spark.sql.rf._
 import org.apache.spark.sql.types.{DataType, StringType}
 import org.apache.spark.sql.{Column, TypedColumn}
 import org.apache.spark.unsafe.types.UTF8String
 import org.locationtech.rasterframes.ref.RasterSource
-
 
 /**
  * Catalyst generator to convert a geotiff download URL into a series of rows